+++ superset/assets/javascripts/explore/components/ExploreViewContainer.jsx
@@ -8,12 +8,15 @@ import ControlPanelsContainer from './ControlPanelsContainer';
 import SaveModal from './SaveModal';
 import QueryAndSaveBtns from './QueryAndSaveBtns';
 import { getExploreUrl } from '../exploreUtils';
-import * as actions from '../actions/exploreActions';
 import { getFormDataFromControls } from '../stores/store';
+import * as exploreActions from '../actions/exploreActions';
+import * as saveModalActions from '../actions/saveModalActions';
+import * as chartActions from '../actions/chartActions';
 
 const propTypes = {
   actions: PropTypes.object.isRequired,
   datasource_type: PropTypes.string.isRequired,
+  isDatasourceMetaLoading: PropTypes.bool.isRequired,
   chartStatus: PropTypes.string,
   controls: PropTypes.object.isRequired,
   forcedHeight: PropTypes.string,
@@ -84,7 +87,6 @@ class ExploreViewContainer extends React.Component {
     const navHeight = this.props.standalone ? 0 : 90;
     return `${window.innerHeight - navHeight}px`;
   }
-
 
   triggerQueryIfNeeded() {
     if (this.props.triggerQuery && !this.hasErrors()) {
@@ -172,7 +174,9 @@ class ExploreViewContainer extends React.Component {
             <ControlPanelsContainer
               actions={this.props.actions}
               form_data={this.props.form_data}
+              controls={this.props.controls}
               datasource_type={this.props.datasource_type}
+              isDatasourceMetaLoading={this.props.isDatasourceMetaLoading}
             />
           </div>
           <div className="col-sm-8">
@@ -186,21 +190,23 @@ class ExploreViewContainer extends React.Component {
 
 ExploreViewContainer.propTypes = propTypes;
 
-function mapStateToProps(state) {
-  const form_data = getFormDataFromControls(state.controls);
+function mapStateToProps({ explore, chart }) {
+  const form_data = getFormDataFromControls(explore.controls);
   return {
-    chartStatus: state.chartStatus,
-    datasource_type: state.datasource.type,
-    controls: state.controls,
+    isDatasourceMetaLoading: explore.isDatasourceMetaLoading,
+    datasource_type: explore.datasource.type,
+    controls: explore.controls,
     form_data,
-    standalone: state.standalone,
-    triggerQuery: state.triggerQuery,
-    forcedHeight: state.forced_height,
-    queryRequest: state.queryRequest,
+    standalone: explore.standalone,
+    triggerQuery: explore.triggerQuery,
+    forcedHeight: explore.forced_height,
+    queryRequest: chart.queryRequest,
+    chartStatus: chart.chartStatus,
   };
 }
 
 function mapDispatchToProps(dispatch) {
+  const actions = Object.assign({}, exploreActions, saveModalActions, chartActions);
   return {
     actions: bindActionCreators(actions, dispatch),
   };
